Moved "algorithms" folder out of the "entities" folder
