+++ ini-rw/algorithms/IniKeyAlgorithms.cpp
